% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-distace.R
\name{calc_distace}
\alias{calc_distace}
\title{calc_distance}
\usage{
calc_distace(org, comparison.orgs, weights)
}
\arguments{
\item{org}{output from \code{get_org_values()}}

\item{comparison.orgs}{output from \code{dat_filtering()}}

\item{weights}{output from \code{get_weights()}}
}
\value{
input of \code{comparison.orgs} with appended distances for each nonprofit
from the input organization
}
\description{
Calculate distances between one organization and all other organizations
Threshold still needs to be added
}
