% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-distace.R
\name{calc_distace_r}
\alias{calc_distace_r}
\title{calc_distance_r}
\format{
\subsection{\code{calc_distance_r}}{
}
}
\usage{
calc_distace_r(org, dat.filtered, weights)
}
\value{
dat.filtered with appended distances
}
\description{
Calculate distances for regular organizations
}
