% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-filtering.R
\name{dat_filtering}
\alias{dat_filtering}
\title{dat_filtering}
\format{
\subsection{\code{dat_filtering}}{
}
}
\usage{
dat_filtering(
  BroadCategroy = NA,
  MajorGroup = NA,
  Decile = NA,
  CommonCode = NA,
  University = NA,
  Hostipal = NA,
  Location = NA,
  State = NA,
  TotalExpenses = c(0, Inf)
)
}
\arguments{
\item{BroadCategroy}{vector of broad categories you wish to include in returned data set}

\item{MajorGroup}{vector of major categories you wish to include in returned data set}

\item{Decile}{vector of numbers from 2 to 10 of decile codes you want to include, exists within MajorGroup}

\item{CommonCode}{"speciality" or "regular" , do you want to include common code orgainzations?}

\item{University}{TRUE of FALSE, do you want to include universities}

\item{Hostipal}{TRUE of FALSE, do you want to include hospitals}

\item{Location}{"metro" or "rural" or "both"}

\item{State}{vector of 2 letter state abbreviations to be included}
}
\description{
STILL NEED TO EDIT
STILL NEED TO ADD EXAMPLE
}
\details{
Filtering \code{nonprofits} with given search criteria
}
