% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dollarize}
\alias{dollarize}
\title{Print a number with U.S. dollar sign and commas}
\usage{
dollarize(x)
}
\arguments{
\item{x}{Numeric value}
}
\value{
\code{x} printed in a character string in U.S. dollar format.
}
\description{
\code{dollarize(x)} inputs a numeric values and returns that number in U.S.
dollar format rounded to the nearest integer.
}
\examples{
dollarize(10)
dollarize(321.53)
dollarize(12000)

}
