% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EIN.filtering}
\alias{EIN.filtering}
\title{EIN.filtering}
\format{
\subsection{\code{EIN_filtering}}{

A data frame with 12,847 rows and 21 columns:
\describe{
\item{EIN}{Employer Identification Number}
\item{ntee}{Orignal NTEE Code}
\item{broad.category}{1-12 Broad Category}
\item{major.group}{A-Z Major Group}
\item{type.org}{"R" for regular organization (two digit value of 20-99) of "S" for specality organization (two digit value of 01-19)}
\item{two.digit}{Orignal decile and centile values of NTEE code}
\item{two.digit.s}{For regular organizations, this is just the two.digit value.
For specialty organizations, if there is a further categorization available this is the 3rd and 4th digit of the NTEE code,
if there is not a further categorization this is just the two digit value. }
\item{tens}{Tens place of the two.digit.s value.}
\item{ones}{Ones place of the two.digit.s value}
\item{us.state}{TRUE if organization is in a U.S. state. FALSE if orgainization if a U.S. territory.}
\item{univ}{TRUE if organization is a university. FALSE otherwise. }
\item{hosp}{TRUE if organization is a hospital. FALSE otherwise. }
\item{total.expense"}{Total anual expenses of organization reported to the IRS.}
\item{state}{2 letter state abbreviation of locaiton of organization.}
\item{location.type}{"metro" or "rural" for location type of organization. }
}
}
}
\usage{
EIN.filtering
}
\description{
data used for filtering in\code{dat-filtering}.
Use EIN to match with \code{nonprofits} data set.
}
\keyword{datasets}
