% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geo-dist.R
\name{get_geo_dist}
\alias{get_geo_dist}
\title{get_geo_dist}
\format{
\subsection{\code{get_geo_dist}}{
}
}
\usage{
get_geo_dist(org1, org2, geo.weights)
}
\arguments{
\item{org1}{output of \code{get_org_values()} for first org}

\item{org2}{output of \code{get_org_values()} for second org}

\item{geo.weights}{geo.weights from output of \code{get_weights()}}
}
\value{
Geographic distance of two orgs
}
\description{
Calculate distances for regular organizations
}
