% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geo-weights.R
\name{get_geo_weights}
\alias{get_geo_weights}
\title{Get weights for geographic distance}
\usage{
get_geo_weights()
}
\value{
Data frame with 2 levels and 2 weights needed to calculated geographic distances.
}
\description{
At the moment, this is the same for all organizations

See vignette for how these values are determined.

Level one represents
}
