% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-new-ntee.R
\name{get_new_ntee}
\alias{get_new_ntee}
\title{get_new_ntee}
\format{
\subsection{\code{get_new_ntee}}{
}
}
\usage{
get_new_ntee(old.code = "B20")
}
\value{
Table with new ntee codes
}
\description{
Input old ntee codes, get new ntee codes
right now this only works for orgs with the 1 letter + 2 digit format,
eventually, want this to work with orgs that have 1 letter + 4 digits
}
