% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-org-values.R
\name{get_org_values}
\alias{get_org_values}
\title{Get Organization Values}
\format{
\code{get_org_values()}
}
\usage{
get_org_values(
  state = "CA",
  location.type = "rural",
  total.expense = 1e+06,
  ntee = "A20"
)
}
\arguments{
\item{state}{Two letter abbreviation of which state your organization is located in. See \code{state.abb52} for options.}

\item{location.type}{either "rural" or "metro" for the type of city your organization is located in.}

\item{total.expense}{a number of the annual total expenses of your organization in U.S. dollars.}

\item{ntee}{a character string of the NTEE code your organization belongs to. See https://nccs.urban.org/publication/irs-activity-codes for a list of all possible codes.}
}
\value{
a list of characteristics about your organization
}
\description{
input basic information about an organization and
return a list of more detailed information.
}
\examples{
get_org_values(state = "NY", location.type = "rural", total.expense = 50000, ntee = "K93")
get_org_values(state = "DC", location.type = "metro", total.expense = 210000, ntee = "A0161")

}
