% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-r-mission-dist.R
\name{get_r_mission_dist}
\alias{get_r_mission_dist}
\title{get_r_mission_dist}
\format{
\subsection{\code{get_r_mission_dist}}{
}
}
\usage{
get_r_mission_dist(org1, org2, r.mission.weights)
}
\arguments{
\item{org1}{output of \code{get_org_values()} for first org}

\item{org2}{output of \code{get_org_values()} for second org}

\item{r.mission.weights}{mission.weights from output of \code{get_weights()}}
}
\value{
mission distance of two orgs
}
\description{
Calculate mission distances for regular organizations
}
