% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-r-mission-weights.R
\name{get_r_mission_weights}
\alias{get_r_mission_weights}
\title{Get weights for regular mission distance}
\usage{
get_r_mission_weights()
}
\description{
At the moment, this is the same for all regular organizations
This is currently a dummy weight set.
This will need to be updated when we decide on the final weight set
}
