% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-s-mission-dist.R
\name{get_s_mission_dist}
\alias{get_s_mission_dist}
\title{get_s_mission_dist}
\format{
\subsection{\code{get_s_mission_dist}}{
}
}
\usage{
get_s_mission_dist(org1, org2, s.mission.weights)
}
\arguments{
\item{org1}{output of \code{get_org_values()} for first org}

\item{org2}{output of \code{get_org_values()} for second org}

\item{s.mission.weights}{mission.weights from output of \code{get_weights()}}
}
\value{
mission distance distance of two orgs
}
\description{
Calculate mission distances for specialty organizations
}
