% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-s-mission-weights.R
\name{get_s_mission_weights}
\alias{get_s_mission_weights}
\title{Get weights for specality mission distance}
\usage{
get_s_mission_weights()
}
\value{
Data frame with 3 levels and 3 weights needed to calculated geographic distances.
}
\description{
At the moment, this is the same for all specality organizations
This is currently a dummy weight set.
This will need to be updated when we decide on the final weight set
}
