% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-salary-range.R
\name{get_salary_range}
\alias{get_salary_range}
\title{Get Salary Range}
\usage{
get_salary_range(samp, point.value)
}
\arguments{
\item{samp}{output from \code{select_sample()}}

\item{point.value}{weighted average salary range from samp}
}
\description{
function to get the salary range
}
