% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-weights.R
\name{get_weights}
\alias{get_weights}
\title{Obtain appropriate weight set}
\usage{
get_weights()
}
\value{
A list with 2 data frames. \code{geo} is the geographic distance weights,
and \code{mission} is for the mission distance weights.
}
\description{
Obtain weight set needed to calculate distances between nonprofits. This is
a wrapper function for \code{get_geo_weights} and \code{get_mission_weights}. This is
function is used in \code{calc_dist}. \code{get_weights} is written as a separate function
from \code{calc_dist} to allow the user to easily change the weights to their own
desired values.

See Appraisal Process Vignette for detailed information about these values.
}
