% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonprofits}
\alias{nonprofits}
\title{nonprofits}
\format{
\subsection{\code{nonprofits}}{

A data frame with 12,847 rows and 21 columns:
\describe{
\item{EIN}{nonprofit IRS employer identification number}
\item{form.year}{IRS Filing year}
\item{name}{Name of nonprofit}
\item{ntee}{NTEE code of nonprofit. See references for details }
\item{univ}{TRUE or FALSE, is the nonprofit a university?}
\item{hosp}{TRUE or FALSE, is the nonprofit a hospital?}
\item{total.expense, TotalEmployee, GrossReceipts, TotalAssests}{respective items reported to the IRS for the filing year}
\item{ceo.compensation}{Total CEO compensation}
\item{gender}{Imputed gender of the CEO, see references details.}
\item{state}{State the nonprofit is in }
\item{zip5}{5 didget zip code of the nonprofit}
\item{location.type}{"metro" or "rural", what type of city the nonprofit is located in}
}
}
}
\usage{
nonprofits
}
\description{
STILL NEED TO EDIT

All non profits available for comparison.
Each non profit who filed with the IRS is in here once, with their most recent filing
Use \code{EIN} to match with \code{EIN_filtering} data set.
}
\keyword{datasets}
