% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntee.crosswalk}
\alias{ntee.crosswalk}
\title{NTEE Codes Crosswalk}
\format{
\subsection{\code{nonprofits.detailed}}{

\describe{
\item{ntee}{Orignal NTEE Code}
\item{new.code}{New NTEE Code. See vignette for how this value is constructed.}
\item{type.org}{Two letter value specifying the type of organization. First part of \code{new.code}. See vignette for detials.}
\item{broad.category}{Three letter value specifying broad category of the organization. Second part of the \code{new.code}. See vignette for details.}
\item{major.group}{A-Z Major Group (3rd to last character of \code{new.code}).}
\item{further.category}{4th and 5th charcaters of \code{old.code} if available, blank otherwise.}
\item{division.subdivision}{Division and Subdivision. See vignette for how this value is determined.}
\item{univ}{TRUE if organization is a university. FALSE otherwise. }
\item{hosp}{TRUE if organization is a hospital. FALSE otherwise. }
}
}
}
\usage{
ntee.crosswalk
}
\description{
Crosswalk between original NTEE codes and the disaggregated version for
calculating distances between two mission codes.
See NTEE Codes Vignette for details on how these values are generated.
}
\details{
Data from   https://github.com/Nonprofit-Open-Data-Collective/mission-taxonomies.
}
\keyword{datasets}
