% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntee.orig}
\alias{ntee.orig}
\title{ntee.orig}
\format{
\subsection{\code{ntee.orig}}{

A data frame with 681 rows and 3 columns:
\describe{
\item{ntee}{If \code{ntee} is one letter, the row describes a major group. If \code{ntee} is one letter + two digits, then the row describes an NTEE code.}
\item{description}{Short description of \code{ntee}.}
\item{definition}{Longer definition of \code{ntee}. }
}
}
}
\usage{
ntee.orig
}
\description{
Original NTEE codes and major groups listed at
https://nccs.urban.org/publication/irs-activity-codes.
}
\keyword{datasets}
