% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.abb52}
\alias{state.abb52}
\title{State abbreviations}
\format{
\subsection{\code{state_abb}}{

A vector with 52 elements
}
}
\usage{
state.abb52
}
\description{
Two letter abbreviations of all states plus Washington D.C. and Puerto Rico.
}
\keyword{datasets}
