% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.dist.mat}
\alias{state.dist.mat}
\title{State Distance Matrix}
\format{
\subsection{\code{state.dist.mat}}{
\itemize{
\item the (i, j) entry is the path length from state i to state j.
\item the diagonal is 0
}
}
}
\usage{
state.dist.mat
}
\description{
A 52-by-52 matrix of distances between all US states (and D.C. and PR)
Distance is calculated by creating a network of U.S. states as follows:
\enumerate{
\item Turn all states into vertices and all boarders into edges.
\item Add edges between Alaska and Washington, Hawaii and California, Puerto Rico and Florida.
\item Distance between two states is the path length between them.
}
}
\details{
Equivalently, this is the number of states you need to go through to drive from state A to state B,
with the added boarders of Alaska/Washington, Hawaii/California, and Puerto Rico/Florida.
}
\keyword{datasets}
