% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_abb.R
\docType{data}
\name{state_abb}
\alias{state_abb}
\title{state_abb}
\format{
\subsection{\code{state_abb}}{

A vector with 52 elements
}
}
\usage{
state_abb
}
\description{
Two letter abbreviations of all states plus D.C. and Puerto Rico.
}
\keyword{datasets}
