% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-filtering.R
\name{dat_filtering}
\alias{dat_filtering}
\title{dat_filtering}
\format{
\subsection{\code{dat_filtering}}{
}
}
\usage{
dat_filtering(
  broad.category = NA,
  major.group = NA,
  decile = NA,
  common.code = NA,
  university = NA,
  hospital = NA,
  location = NA,
  state = NA,
  total.expenses = c(0, Inf)
)
}
\arguments{
\item{broad.category}{vector of broad categories you wish to include in returned data set}

\item{major.group}{vector of major categories you wish to include in returned data set}

\item{decile}{vector of numbers from 2 to 10 of decile codes you want to include, exists within major.group}

\item{common.code}{"speciality" or "regular" , do you want to include common code orgainzations?}

\item{university}{TRUE of FALSE, do you want to include universities}

\item{hospital}{TRUE of FALSE, do you want to include hospitals}

\item{location}{"metro" or "rural" or "both"}

\item{state}{vector of 2 letter state abbreviations to be included}
}
\description{
STILL NEED TO EDIT
STILL NEED TO ADD EXAMPLE
}
\details{
Filtering \code{nonprofits} with given search criteria
}
