% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-filtering.R
\name{dat_filtering}
\alias{dat_filtering}
\title{dat_filtering}
\format{
\subsection{\code{dat_filtering}}{
}
}
\usage{
dat_filtering(
  broad.category = 1:12,
  major.group = base::LETTERS,
  tens = 2:9,
  type.org = "regular",
  univ = FALSE,
  hosp = FALSE,
  location.type = "both",
  state = state.abb52,
  total.expense = c(0, Inf)
)
}
\arguments{
\item{broad.category}{vector of broad categories you wish to include in returned data set}

\item{major.group}{vector of major categories you wish to include in returned data set}

\item{tens}{vector of numbers from 2 to 10 of decile codes you want to include, exists within major.group.}

\item{type.org}{"specialty" or "regular" or "both".
"specialty" to only include specialty (common code) organization.
"regular" to only include regular organizations.
"both" to include both.}

\item{univ}{TRUE of FALSE. Are universities to be included?}

\item{hosp}{TRUE of FALSE, Are hospitals to be included?}

\item{location.type}{"metro" or "rural" or "both" or "all".
"metro" only includes organizations with metropolitan locations type.
"rural" only includes organizations with rural locations type.
"both" includes both metropolitan and rural organizations but excludes organizations with NA values in the location type.
"all" includes organizations with metropolitan, rural, and missing location types.}

\item{state}{vector of 2 letter state abbreviations to be included}

\item{total.expense}{vector of c(min,max) of range of total expenses to be included}
}
\value{
All entries in \code{EIN_filtering} data set that match the filtering criteria
}
\description{
STILL NEED TO ADD EXAMPLE
}
\details{
STILL NEED TO EDIT DESCRIPtion
Filtering \code{EIN_filtering} with given search criteria.
EIN's that make it through this filtering, will be used to filter \code{nonprofits} for final comparison sets.
Check https://nccs.urban.org/project/national-taxonomy-exempt-entities-ntee-codes for NTEE code explanations.
}
\examples{
# all non-univ educational nonprofits in Kansas, Nebraska, Iowa, and Missouri  
dat_filtering(
  broad.category = 2, 
  major.group = "B", 
  tens = 2:9, 
  type.org = "regular", 
  univ = FALSE,
  hosp = FALSE, 
  location.type = "both", 
  state = c("KS", "NE", "IA", "MO") ,
  total.expense = c(0, Inf))

}
