% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-filtering.R
\name{dat_filtering}
\alias{dat_filtering}
\title{dat_filtering}
\format{
\subsection{\code{dat_filtering}}{
}
}
\usage{
dat_filtering(
  type.org = "RG",
  broad.category = c("ART", "EDU", "ENV", "HEL", "HMS", "IFA", "PSB", "REL", "MMB",
    "UNU", "UNI", "HOS"),
  major.group = base::LETTERS,
  division = NA,
  subdivision = NA,
  univ = FALSE,
  hosp = FALSE,
  location.type = c("urban", "suburban", "town", "rural"),
  state = state.abb52,
  total.expense = c(0, Inf)
)
}
\arguments{
\item{type.org}{vector of the types of organization you want to include. Options are  RG, AA, MT, PA, RP, MS, MM, and/or NS.}

\item{broad.category}{vector of broad categories you wish to include in returned data set Options are ART, EDU, ENV, HEL, HMS, IFA, PSB, REL, MMB, UNU, UNI, and/or HOS}

\item{major.group}{vector of major groups you wish to include in returned data set. Options are A-Z.}

\item{division}{vector of divisions you wish to include. Divisions exist entirely inside major groups. We suggest you do not use this parameter if you have more than one item in \code{major.group}. Options are 0, 2, 3, ..., 9 (1 is not an option.}

\item{subdivision}{vector of subdivision you wish to include. Subdivisions exist entirely inside divisions. We suggest you do not use this parameter if you have more than one item in \code{division}. Options are 0 - 9.}

\item{univ}{TRUE of FALSE. Are universities to be included?}

\item{hosp}{TRUE of FALSE, Are hospitals to be included?}

\item{location.type}{vector of "metro", "suburban", "town", and/or "rural" for which city types to include}

\item{state}{vector of 2 letter state abbreviations to be included}

\item{total.expense}{vector of c(min,max) of range of total expenses to be included

Any parameter you want to include all values for, you can either list all possible values, or assign that parameter to NA.
Default parameters are set to include all regular organizations and exclude all specality organizations.}
}
\value{
All entries in \code{nonprofits.detailed} data set that match the filtering criteria
}
\description{
STILL NEED TO ADD EXAMPLE
}
\details{
STILL NEED TO EDIT DESCRIPtion
Filtering \code{nonprofits.detailed} with given search criteria.
EIN's that make it through this filtering, will be used to filter \code{nonprofits} for final comparison sets.
Check https://nccs.urban.org/project/national-taxonomy-exempt-entities-ntee-codes for NTEE code explanations.
}
\examples{
# all non-university educational nonprofits in Kansas, Nebraska, Iowa, and Missouri.
dat_filtering(
  broad.category = "EDU", 
  major.group = "B", 
  type.org = "RG", 
  univ = FALSE,
  hosp = FALSE, 
  location.type = c("urban", "suburban", "town", "rural"), 
  state = c("KS", "NE", "IA", "MO") ,
  total.expense = c(0, Inf))

}
