% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-appraisal.R
\name{get_appraisal}
\alias{get_appraisal}
\title{Get compensation Appraisal}
\usage{
get_appraisal(org, search.criteria)
}
\arguments{
\item{org}{output from \code{get_org_values()}}

\item{search.criteria}{a list with the following items:
\itemize{
\item type.org vector of the types of organization you want to include. Options are  RG, AA, MT, PA, RP, MS, MM, and/or NS.
\item broad.category vector of broad categories you wish to include in returned data set Options are ART, EDU, ENV, HEL, HMS, IFA, PSB, REL, MMB, UNU, UNI, and/or HOS
\item major.group vector of major groups you wish to include in returned data set. Options are A-Z.
\item division vector of divisions you wish to include. Divisions exist entirely inside major groups. We suggest you do not use this parameter if you have more than one item in \code{major.group}. Options are 0, 2, 3, ..., 9 (1 is not an option.
\item subdivision vector of subdivision you wish to include. Subdivisions exist entirely inside divisions. We suggest you do not use this parameter if you have more than one item in \code{division}. Options are 0 - 9.
\item univ TRUE of FALSE. Are universities to be included?
\item hosp TRUE of FALSE, Are hospitals to be included?
\item location.type vector of "metro", "suburban", "town", and/or "rural" for which city types to include
\item state vector of 2 letter state abbreviations to be included
\item total.expense vector of c(min,max) of range of total expenses to be included
}}
}
\value{
A list with 3 objects
\enumerate{
\item \code{suggested.salary}: a numeric of the suggested CEO compensation for the input organization
\item \code{suggested.range}: vector of minimum and maximum suggested salary range
\item \code{reference.set}: a data frame of all nonprofits used to calculated the suggested
CEO compensation with the following variables:
}
\itemize{
\item \code{EIN}: IRS Employer Identification Number
\item \code{form.year}: IRS filing year from which this nonprofits information was obtained
\item \code{name}: Name of the nonprofit
\item \code{total.employee}: Total number of employees at the nonprofit
\item \code{gross.receipts}: Gross receipts reported for the year
\item \code{total.assests}: Total assets reported for the year
\item \code{total.expense}: Total expenses reported for the year
\item \code{ceo.compensation}: Total CEO compensation reported for the year
\item \code{gender}: Imputed gender of the CEO
\item \code{zip5}: 5 digit zip code of where the nonprofit is located
\item \code{state}: Two letter abbreviation of the state where the nonprofit is located
\item \code{location.type}: Either "metro" or "rural" for type of location the nonprofit is in
\item \code{ntee}: Original ntee code
\item \code{broad.category}, \code{major.group}, \code{type.org}, \code{two.digit}, \code{two.digit.s}, \code{tens}, \code{ones}, \code{us.state}, \code{univ}, \code{hosp}: Parts of the dissagregated NTEE code. See ... for details.
\item \code{log.expense.dist}: Total Expense distance between the nonprofit and the reference organization
\item \code{mission.dist}: Mission distance between the nonprofit and the reference organization
\item \code{geo.dist}: Geographic distance between the nonprofit and the reference organization
\item \code{total.dist}: Total distance between the nonprofit and the reference organization
\item \code{rank}: Ranking of all nonprofits that match the reference set from closest to farthest from the reference organization.
\item \code{dist.std}: standardized total distance
\item \code{weight}: weight used to calculated weighted average of \code{suggested.salary}
\item \code{residual.percent}: the residual of that observation as a percent of the expected salary.
\item \code{fitted.values}: is \code{suggested.salary}(1+\code{residual.percent})
}

See ... Vignette for details on how these values are calculated.
}
\description{
Wrapper function for generating entire compensation appraisal
}
\examples{

\dontrun{
# Step 1: Create list of organization characteristics 
org <- get_org_values(
         state = "FL",
         location.type = "metro",
         total.expense = 1000000,
         ntee = "P20")
search.criteria <-
  list(
    type.org = "RG",
    broad.category = "HMS",
    major.group = c("M", "N", "O", "P"),
    division = NA,
    subdivision = NA,
    univ = FALSE,
    hosp = FALSE,
    location.type = c("metro", "suburban"),
    state = c("FL", "PR", "GA", "SC", "MS", "TN", "AL"),
    total.expense = c(100000, 10000000)
  )
       
#Step 3: Get appraisal
appraisal <- get_appraisal(org, search.criteria) 

appraisal$suggested.salary
appraisal$suggested.range
reference.set <- appraisal$reference.set
View(reference.set)
}
}
