% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-appraisal.R
\name{get_appraisal}
\alias{get_appraisal}
\title{Get compensation Appraisal}
\usage{
get_appraisal(org.characteristics, search.criteria)
}
\arguments{
\item{org.characteristics}{output from \code{get_org_values()}}

\item{search}{criteria a list with ...... NEED TO UPDATE}
}
\description{
Wrapper function for generating entire compensation appraisal
}
