% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geo-dist.R
\name{get_geo_dist}
\alias{get_geo_dist}
\title{Calculate geographic distance between two nonprofits}
\usage{
get_geo_dist(org1, org2, geo.weights)
}
\arguments{
\item{org1}{output of \code{get_org_values()} for first org}

\item{org2}{output of \code{get_org_values()} for second org}

\item{geo.weights}{geo.weights from output of \code{get_weights()}}
}
\value{
Numeric value of the geographic distance of two nonprofits.
}
\description{
Calculate geographic distance between two nonprofits. Primarily as an internal
function in \code{calc_distance}, but can be used on its own. See Appraisal Process
Vignette for detailed information on how this value is calculated.
}
\examples{
org1 <- get_org_values(
  state = "IL", 
  location.type = "metro",
  total.expense = 500000,
  ntee = "A20")
  
org2 <- get_org_values(
  state = "OH", 
  location.type = "suburban",
  total.expense = 300000,
  ntee = "A20")
  
geo.weights <- get_geo_weights()
  
get_geo_dist(org1, org2, geo.weights)
  
}
