% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geo-dist.R
\name{get_geo_dist}
\alias{get_geo_dist}
\title{get_geo_dist}
\format{
\subsection{\code{get_geo_dist}}{
}
}
\usage{
get_geo_dist(org1, org2, geo.weights)
}
\value{
Geographic distance of two orgs
}
\description{
Calculate distances for regular organizations
}
