% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geo-weights.R
\name{get_geo_weights}
\alias{get_geo_weights}
\title{Get weights for geographic distance}
\usage{
get_geo_weights()
}
\description{
At the moment, this is the same for all organizations
This is currently a dummy weight set.
This will need to be updated when we decide on the final weight set
}
