% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geo-weights.R
\name{get_geo_weights}
\alias{get_geo_weights}
\title{Get weights for geographic distance}
\usage{
get_geo_weights()
}
\value{
Data frame with 2 levels and 2 weights needed to calculated geographic distances.
}
\description{
Weights for geographic distance. See Appraisal Process Vignette for the interpretation
and usage of these values.

This is primarily used as an internal function for \code{get_geo_dist}. \code{get_geo_weights}
is written as a separate function from \code{get_geo_dist} to allow the user to easily
change the weights to their own desired values.
}
