% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mission-dist.R
\name{get_mission_dist}
\alias{get_mission_dist}
\title{Calculate mission distance between two nonprofits}
\format{
\subsection{\code{get_mission_dist}}{
}
}
\usage{
get_mission_dist(org1, org2, mission.weights)
}
\arguments{
\item{org1}{output of \code{get_org_values()} for first org}

\item{org2}{output of \code{get_org_values()} for second org}

\item{mission.weights}{mission.weights from output of \code{get_weights()}}
}
\value{
Numeric value of the geographic distance of two nonprofits.
}
\description{
Calculate mission distance between two nonprofits. Primarily as an internal
function in \code{calc_distance}, but can be used on its own. See Appraisal Process
Vignette for detailed information on how this value is calculated.
}
\examples{
org1 <- get_org_values(
  state = "IL", 
  location.type = "metro",
  total.expense = 500000,
  ntee = "Q1120")
  
org2 <- get_org_values(
  state = "OH", 
  location.type = "suburban",
  total.expense = 300000,
  ntee = "A35")
  
mission.weights <- get_mission_weights()
  
get_mission_dist(org1, org2, mission.weights)

}
