% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mission-weights.R
\name{get_mission_weights}
\alias{get_mission_weights}
\title{Get weights for regular mission distance}
\usage{
get_mission_weights()
}
\value{
Data frame with 5 levels and 5 weights needed to calculated mission distances.
}
\description{
Weights for mission distance
}
