% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-new-ntee.R
\name{get_new_ntee}
\alias{get_new_ntee}
\title{Obtain disaggregated NTEE code}
\usage{
get_new_ntee(old.code = "B20")
}
\arguments{
\item{old.code}{character string of original NTEE code. See above NTEE Codes Vignette for acceptable NTEE Codes.}
}
\value{
A list with \code{new.code}, the new disaggregated code, and other related values.
See NTEE Codes Vignette  on how these values are calculated.
}
\description{
Input original NTEE code and return disaggregated NTEE code.

All known NTEE codes and their respective disaggregated codes are stored
in \code{ntee-crosswalk} data set. All NTEE codes without a further categroization
can be found at https://nccs.urban.org/publication/irs-activity-codes.

In this appraisal process, this function is used to calculated distances between
mission codes.
}
\examples{
get_new_ntee("H96")
get_new_ntee("R02")
get_new_ntee("R0226")
}
