% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-new-ntee.R
\name{get_new_ntee}
\alias{get_new_ntee}
\title{Disaggregate NTEE Code}
\format{
\subsection{\code{get_new_ntee}}{
}
}
\usage{
get_new_ntee(old.code = "B20")
}
\arguments{
\item{old.code}{character string of original NTEE code}
}
\value{
A list with \code{new.code}, thenew disaggregated code, and other related values.
See ... Vignette on how these values are calculated.
All known NTEE codes and their respective disaggregated codes are stored
in \code{ntee-crosswalk} data set.
}
\description{
Input original NTEE code and return disaggregated NTEE code.

A list of all available NTEE codes can be found in the \code{ntee.crosswalk}
data set and at https://nccs.urban.org/publication/irs-activity-codes.

Used to calculated distances between mission codes.
}
\examples{
get_new_ntee("H96")
get_new_ntee("R02")
get_new_ntee("R0226")
}
