% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-new-ntee.R
\name{get_new_ntee}
\alias{get_new_ntee}
\title{get_new_ntee}
\format{
\subsection{\code{get_new_ntee}}{
}
}
\usage{
get_new_ntee(old.code = "B20")
}
\value{
Table with new ntee codes
}
\description{
Input original NTEE code and return disaggregated NTEE code.
}
\details{
Used to calculated distances between mission codes.
}
