% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-r-mission-dist.R
\name{get_r_mission_dist}
\alias{get_r_mission_dist}
\title{get_r_mission_dist}
\format{
\subsection{\code{get_r_mission_dist}}{
}
}
\usage{
get_r_mission_dist(org1, org2, r.mission.weights)
}
\value{
Geographic distance of two orgs
}
\description{
Calculate mission distances for regular organizations
}
