% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-salary-range.R
\name{get_salary_range}
\alias{get_salary_range}
\title{Get Salary Range}
\usage{
get_salary_range(samp, point.value)
}
\arguments{
\item{samp}{output from \code{select_sample()}}

\item{point.value}{weighted average salary range from samp}
}
\value{
A list with
\enumerate{
\item \code{suggested.range}: vector of minimum and maximum suggested salary range
\item \code{samp}: the input data frame \code{samp} with two new columns. \code{residual.percent} is the
residual of that observation as a percent of the expected salary. \code{fitted.values} weighted
average of all CEO compensation multiplied by (1+\code{residual.percent})
}

See ... Vignette for detailed explanation on how these values are calculated
}
\description{
function to get the salary range
}
