% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-weights.R
\name{get_weights}
\alias{get_weights}
\title{Obtain appropriate weight set}
\usage{
get_weights(org = get_org_values())
}
\arguments{
\item{org}{list output from \code{get_org_values()}}
}
\description{
Input user org characteristics, output correct weight set to be used for distance
}
