% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonprofits}
\alias{nonprofits}
\title{nonprofits}
\format{
\subsection{\code{nonprofits}}{

A data frame with 12,847 rows and 18 columns.
Each row is a unique nonprofit, and the columns are its respective attributes:
\describe{
\item{EIN}{nonprofit IRS employer identification number}
\item{form.year}{IRS Filing year}
\item{name}{Name of nonprofit}
\item{url}{URL of ProPublica.org with more detailed information about the nonprofit.}
\item{ntee}{Original NTEE code of nonprofit. See references for details }
\item{new.code}{New disaggregated NTEE code. See vignette for details.}
\item{univ}{TRUE or FALSE, is the nonprofit a university?}
\item{hosp}{TRUE or FALSE, is the nonprofit a hospital?}
\item{total.expense, total.employee, gross.reciepts, total.assests}{respective items reported to the IRS for the filing year}
\item{ceo.compensation}{Total CEO compensation}
\item{gender}{Imputed gender of the CEO, see references for details.}
\item{state}{State the nonprofit is located in }
\item{zip5}{5 didget zip code the nonprofit is located in}
\item{location.type}{"metro", "subruban", "town", or "rural", what type of city the nonprofit is located in}
}
}
}
\usage{
nonprofits
}
\description{
All nonprofits that electronically filled with the IRS from 2009-2019.
Each nonprofit is listed in this data set exactly once at their most recent
file time. See vignette for detailed description on how this data set was
created.

This is all non profits available for use when generating an appraisal.

Use \code{EIN} to match with \code{nonprofits.detailed} data set for more detailed
information about each nonprofit.
}
\keyword{datasets}
