% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntee.crosswalk}
\alias{ntee.crosswalk}
\title{NTEE Codes Crosswalk}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 655 rows and 12 columns.
}
\usage{
ntee.crosswalk
}
\description{
NTEE Codes Crosswalk
}
\keyword{datasets}
