% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntee.crosswalk}
\alias{ntee.crosswalk}
\title{NTEE Codes Crosswalk}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1086 rows and 10 columns.
}
\usage{
ntee.crosswalk
}
\description{
Crosswalk between original NTEE codes and the disaggregated version for calculating distances between two mission codes.
See ... Vignette for details on how these values are calculated
}
\details{
Data from   https://github.com/Nonprofit-Open-Data-Collective/mission-taxonomies.
}
\keyword{datasets}
