% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-salary.R
\name{predict_salary}
\alias{predict_salary}
\title{Get Salary Appraisal}
\usage{
predict_salary(samp)
}
\arguments{
\item{samp}{output from \code{select_sample()}}
}
\description{
function to predict salary
}
