% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-sample.R
\name{select_sample}
\alias{select_sample}
\title{Select Sample}
\usage{
select_sample(
  org = get_org_values(state = "AL", location.type = "rural", total.expense = 1e+05, ntee
    = "P20"),
  search.criteria = list(type.org = "RG", broad.category = NA, major.group = LETTERS,
    division = NA, subdivision = NA, univ = FALSE, hosp = FALSE, location.type =
    c("rural", "town"), state = state.abb52, total.expense = c(1e+06, 1e+07))
)
}
\arguments{
\item{org}{List output from \code{get_org_values()}}

\item{search.criteria}{A list with the following elements:
\itemize{
\item \code{type.org}: vector of the types of organization you want to include. Options are  RG, AA, MT, PA, RP, MS, MM, and/or NS.
\item \code{broad.category}: vector of broad categories you wish to include in returned data set Options are ART, EDU, ENV, HEL, HMS, IFA, PSB, REL, MMB, UNU, UNI, and/or HOS
\item \code{major.group}: vector of major groups you wish to include in returned data set. Options are A-Z.
\item \code{division}: vector of divisions you wish to include. Divisions exist entirely inside major groups. We suggest you do not use this parameter if you have more than one item in \code{major.group}. Options are 0, 2, 3, ..., 9 (1 is not an option.
\item \code{subdivision}: vector of subdivision you wish to include. Subdivisions exist entirely inside divisions. We suggest you do not use this parameter if you have more than one item in \code{division}. Options are 0 - 9.
\item \code{univ}: TRUE of FALSE. Are universities to be included?
\item \code{hosp}: TRUE of FALSE, Are hospitals to be included?
\item \code{location.type}: vector of "metro", "suburban", "town", and/or "rural" for which city types to include
\item \code{state}: vector of 2 letter state abbreviations to be included
\item \code{total.expense}: vector of \code{c(min,max)} of range of total expenses to be included
}}
}
\value{
A data frame with all nonprofits that match the search criteria.
Each nonprofit has the following variables:
\itemize{
\item \code{EIN}: IRS Employer Identification Number
\item \code{form.year}: IRS filing year from which this nonprofits information was obtained
\item \code{name}: Name of the nonprofit
\item \code{total.employee}: Total number of employees at the nonprofit
\item \code{gross.receipts}: Gross receipts reported for the year
\item \code{total.assests}: Total assets reported for the year
\item \code{total.expense}: Total expenses reported for the year
\item \code{ceo.compensation}: Total CEO compensation reported for the year
\item \code{gender}: Imputed gender of the CEO
\item \code{zip5}: 5 digit zip code of where the nonprofit is located
\item \code{state}: Two letter abbreviation of the state where the nonprofit is located
\item \code{location.type}: Either "metro" or "rural" for type of location the nonprofit is in
\item \code{ntee}: Original NTEE code
\item \code{new.code}: New NTEE code
\item \verb{org.type, }broad.category\verb{, }major.group\verb{, }division\verb{, }subdivision\verb{, }univ\verb{, }hosp`: Parts of the dissagregated NTEE code. See ... for details.
\item \code{expense.dist}: Total Expense distance between the nonprofit and the reference organization
\item \code{mission.dist}: Mission distance between the nonprofit and the reference organization
\item \code{geo.dist}: Geographic distance between the nonprofit and the reference organization
\item \code{total.dist}: Total distance between the nonprofit and the reference organization
\item \code{rank}: Ranking of all nonprofits that match the reference set from closest to farthest from the reference organization.
}
}
\description{
STILL NEED TO EDIT
\code{select_sample} is used to select all other nonprofits that match a specified criteria,
as well as calculate their respective distances from a reference organization.

See ... Vignette for detailed explanation on how distance between nonprofits is calculated.
}
\examples{

input.org <- 
  get_org_values(state = "FL",#
                 location.type = "rural",
                 total.expense = 1.2e6,
                 ntee = "B20")

search.criteria <-
  list(
    type.org = "RG",
    broad.category = "EDU",
    major.group = "B",
    division = 2:9,
    subdivision = NA,
    univ = FALSE,
    hosp = FALSE,
    location.type = NA,
    state = c("FL", "GA", "SC", "MS", "AL", "PR"),
    total.expense = c(1.2e5, 1.2e7)
  )
      
samp <- select_sample(input.org, search.criteria)
}
